fix(rust, python): projection pushdown with union and asof join (#5763)
