+++ polars/polars-lazy/Cargo.toml
@@ -9,6 +9,9 @@ description = "Lazy query engine for the Polars DataFrame library"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[dev-dependencies]
+serde_json = "1"
+
 [dependencies]
 ahash.workspace = true
 bitflags.workspace = true
