+++ polars/polars-lazy/src/physical_plan/executors/union.rs
@@ -89,6 +89,7 @@ impl Executor for UnionExec {
                     })
                     .collect::<PolarsResult<Vec<_>>>()
             });
+            dbg!(&out);
 
             concat_df(out?.iter().flat_map(|dfs| dfs.iter()))
         }
